Module aws_sdk_pinpointsmsvoice::error

Modules

already_exists_exception

See AlreadyExistsException

bad_request_exception

See BadRequestException

internal_service_error_exception

See InternalServiceErrorException

limit_exceeded_exception

See LimitExceededException

not_found_exception

See NotFoundException

too_many_requests_exception

See TooManyRequestsException

Structs

AlreadyExistsException

The resource specified in your request already exists.

BadRequestException

The input you provided is invalid.

CreateConfigurationSetError

CreateConfigurationSetEventDestinationError

DeleteConfigurationSetError

DeleteConfigurationSetEventDestinationError

GetConfigurationSetEventDestinationsError

InternalServiceErrorException

The API encountered an unexpected error and couldn’t complete the request. You might be able to successfully issue the request again in the future.

LimitExceededException

There are too many instances of the specified resource type.

ListConfigurationSetsError

NotFoundException

The resource you attempted to access doesn’t exist.

SendVoiceMessageError

TooManyRequestsException

You’ve issued too many requests to the resource. Wait a few minutes, and then try again.

UpdateConfigurationSetEventDestinationError

Enums

CreateConfigurationSetErrorKind

CreateConfigurationSetEventDestinationErrorKind

DeleteConfigurationSetErrorKind

DeleteConfigurationSetEventDestinationErrorKind

GetConfigurationSetEventDestinationsErrorKind

ListConfigurationSetsErrorKind

SendVoiceMessageErrorKind

UpdateConfigurationSetEventDestinationErrorKind