Module aws_sdk_pinpointsmsvoice::model

Modules

call_instructions_message_type

See CallInstructionsMessageType

cloud_watch_logs_destination

See CloudWatchLogsDestination

event_destination

See EventDestination

event_destination_definition

See EventDestinationDefinition

kinesis_firehose_destination

See KinesisFirehoseDestination

plain_text_message_type

See PlainTextMessageType

sns_destination

See SnsDestination

ssml_message_type

See SsmlMessageType

voice_message_content

See VoiceMessageContent

Structs

CallInstructionsMessageType

An object that defines a message that contains text formatted using Amazon Pinpoint Voice Instructions markup.

CloudWatchLogsDestination

An object that contains information about an event destination that sends data to Amazon CloudWatch Logs.

EventDestination

An object that defines an event destination.

EventDestinationDefinition

An object that defines a single event destination.

KinesisFirehoseDestination

An object that contains information about an event destination that sends data to Amazon Kinesis Data Firehose.

PlainTextMessageType

An object that defines a message that contains unformatted text.

SnsDestination

An object that contains information about an event destination that sends data to Amazon SNS.

SsmlMessageType

An object that defines a message that contains SSML-formatted text.

VoiceMessageContent

An object that contains a voice message and information about the recipient that you want to send it to.

Enums

EventType

The types of events that are sent to the event destination.