Module aws_sdk_pinpointsmsvoice::operation

Structs

CreateConfigurationSet

Create a new configuration set. After you create the configuration set, you can add one or more event destinations to it.

CreateConfigurationSetEventDestination

Create a new event destination in a configuration set.

DeleteConfigurationSet

Deletes an existing configuration set.

DeleteConfigurationSetEventDestination

Deletes an event destination in a configuration set.

GetConfigurationSetEventDestinations

Obtain information about an event destination, including the types of events it reports, the Amazon Resource Name (ARN) of the destination, and the name of the event destination.

ListConfigurationSets

List all of the configuration sets associated with your Amazon Pinpoint account in the current region.

SendVoiceMessage

Create a new voice message and send it to a recipient’s phone number.

UpdateConfigurationSetEventDestination

Update an event destination in a configuration set. An event destination is a location that you publish information about your voice calls to. For example, you can log an event to an Amazon CloudWatch destination when a call fails.