Module aws_sdk_pinpointsmsvoice::output

Modules

create_configuration_set_event_destination_output

See CreateConfigurationSetEventDestinationOutput

create_configuration_set_output

See CreateConfigurationSetOutput

delete_configuration_set_event_destination_output

See DeleteConfigurationSetEventDestinationOutput

delete_configuration_set_output

See DeleteConfigurationSetOutput

get_configuration_set_event_destinations_output

See GetConfigurationSetEventDestinationsOutput

list_configuration_sets_output

See ListConfigurationSetsOutput

send_voice_message_output

See SendVoiceMessageOutput

update_configuration_set_event_destination_output

See UpdateConfigurationSetEventDestinationOutput

Structs

CreateConfigurationSetEventDestinationOutput

An empty object that indicates that the event destination was created successfully.

CreateConfigurationSetOutput

An empty object that indicates that the configuration set was successfully created.

DeleteConfigurationSetEventDestinationOutput

An empty object that indicates that the event destination was deleted successfully.

DeleteConfigurationSetOutput

An empty object that indicates that the configuration set was deleted successfully.

GetConfigurationSetEventDestinationsOutput

An object that contains information about an event destination.

ListConfigurationSetsOutput

An object that contains information about the configuration sets for your account in the current region.

SendVoiceMessageOutput

An object that that contains the Message ID of a Voice message that was sent successfully.

UpdateConfigurationSetEventDestinationOutput

An empty object that indicates that the event destination was updated successfully.